Add nw connectivity task to all playbooks
When using bgp where pods using a network-attachment-definition
it takes some seconds to advertise it on the fabric. As a result
roles which don't have a the connectivity task check will fail.

A rerun of the ansible jobs will not fix this because every time
the deployment job pod gets created it will receive a new ip from
the ctlplane pool.

There were discussions on a more generic way on slack, for
now lets add a connectivity check to the edpm_nodes_validation role
and call it from the playbooks, similar to:
https://github.com/openstack-k8s-operators/edpm-ansible/blob/6615889c5d708df2accb7dafec5aa099946292b8/playbooks/bootstrap.yml#L8

Jira: https://issues.redhat.com/browse/OSPRH-8680

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>